Make AsyncThreadTimedJoin available on other platforms.
On Linux, this uses pthread_timedjoin_np().  For other OSes (not Windows
or Linux), we just fall back to a blocking join :/
